Move storage to its own module
